requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
mediapipe>=0.10.9
opencv-python>=4.9.0
librosa>=0.10.1
gTTS>=2.5.1
numpy>=1.26.0
pyyaml>=6.0.1
tqdm>=4.66.0
scipy>=1.12.0
matplotlib>=3.8.0
jupyter>=1.0.0
ipykernel>=6.29.0